requirements.txt
torch
opencv-python
numpy
Pillow
scikit-learn
requests
